project-1/requirements.txt
psycopg2-binary==2.8.2
sqlalchemy==1.3
